Moved verifyGitTagsDoNotAlreadyExist to common
